requirements.txt
rpi_rf >= 0.9.7
paho-mqtt >= 1.5.0
